Replace timeout generation code with new equation.
Under the old system, the timeout to use was generated through a complicated and somewhat nonsensical
formula that became less tolerant as the computation neared the end.

The new technique is to estimate the amount of time it would take to transfer the file, add in a 25% buffer,
and then also take into account the amount of time it would take to get the worker "back up to speed" weighted
by the percentage of the computation that the worker represents.